third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-whitespace.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    --background-color:{ background-color: green; };
    --width:  {width:100px}  ;
    --height:{height: 200px} ;
    @apply --background-color ;
    @apply/**/--width;
    @apply --height;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).width, "100px");
    assert_equals(getComputedStyle(e).height, "200px");
}, "Whitespace is optional in @apply and custom property sets");
</script>
