third_party/WebKit/LayoutTests/platform/win/fast/css/atapply/at-apply-whitespace-expected.txt

This is a testharness.js-based test.
PASS Whitespace is optional in @apply and custom property sets 
Harness: the test ran to completion.
